backend/requirements.txt
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
boto3==1.34.47
botocore==1.34.47
bs4==0.0.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==3.4.8
Django==4.2.4
django-cors-headers==4.3.1
django-environ==0.10.0
djangorestframework==3.14.0
exceptiongroup==1.2.0
gunicorn==21.2.0
h11==0.14.0
idna==3.6
jmespath==1.0.1
outcome==1.3.0.post0
packaging==23.2
psycopg==3.1.17
psycopg-binary==3.1.17
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.10.0
selenium==4.18.1
sentry-sdk==1.40.5
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
urllib3==2.0.7
whitenoise==6.4.0
wsproto==1.2.0
